merge

+++ src/controllers/FormController.php
@@ -18,6 +18,13 @@ class FormController extends Controller
         }
         return response()->json(["data" => $option]);
     }
+    
+    public function getPreview()
+    {
+        $html = preg_replace("/\r\n|\r|\n/",'',\GritTekno::header(env('menu_app')));
+        return response()->json(["data" => $html]);
+    }
+    
     public function menu($id = null, Request $request)
     {
         try {
@@ -29,48 +36,33 @@ class FormController extends Controller
             }
 
             if ($request->isMethod('delete')) {
-                $model = DataMenu::Find($request->id);
-                $model->name = $model->name.'_'.$request->id;
-                $model->save();
+                $model = DataMenu::Find(Crypt::decryptstring($request->ref));
+                $count=\DB::table('menu')->select(\DB::raw('count(id_parent) as parent_count'))->where('id_parent', $model->id_mnu)->first();
+                if ($count->parent_count > 0) {
+                    return response()->json(["error" => "Delete failed. Child relationship exists!"]);
+                }
                 $model->delete();
                 return 'success';
             }
 
             if ($request->isMethod('post')) {
-                $cek = DataMenu::where('menu_ut',$request->menu_ut)->first();
-                if ($cek != NULL) {
-                    if ($cek->deleted_at != null) {
-                        $cek->deleted_at = NULL;
-                        $cek->save(); 
-                        return 'Success';   
-                    }else {
-                        return response()->json(["error" => "Menu utama sudah terdaftar."]);
-                    }
-                } else {
-                    $model = new DataMenu;
-                    $model->app = env('menu_app');
-                    $model->menu_ut = $request->menu_ut;
-                    $model->id_parent = $request->id_parent ;
-                    $model->divider = $request->divider ? 1 : 0 ;
-                    $model->header = $request->header ? 1 : 0 ;
-                    $model->icon = 'fa '.$request->icon;
-                    $model->routename = $request->routename;
-                    $model->url = $request->routename == '#' ? '#':route($request->routename);
-                    $model->urut = $request->urut ? $request->urut : 100;
-                    $model->save();
-                    return 'Success';
-                }
+                $model = new DataMenu;
+                $model->app = env('menu_app');
+                $model->menu_ut = $request->menu_ut;
+                $model->id_parent = $request->id_parent ;
+                $model->divider = $request->divider ? 1 : 0 ;
+                $model->header = $request->header ? 1 : 0 ;
+                $model->icon = 'fa '.$request->icon;
+                $model->routename = $request->routename;
+                $model->url = $request->routename == '#' ? '#':route($request->routename);
+                $model->urut = $request->urut ? $request->urut : 100;
+                $model->save();
+                return 'Success';
             }
 
             if ($request->isMethod('put')) {
                 $id=Crypt::decryptstring($request->ref);
                 $model = DataMenu::where('id_mnu',$id)->first();
-                if ($model->menu_ut != $request->menu_ut) {
-                    $cek = DataMenu::where('menu_ut',$request->menu_ut)->first();
-                    if ($cek != NULL) {
-                        return response()->json(["error" => "Menu utama sudah terdaftar."]);
-                    }
-                }
                 $model->app = env('menu_app');
                 $model->app = env('menu_app');
                 $model->menu_ut = $request->menu_ut;
